requirements.txt
pymongo==3.11.3
python-telegram-bot
lxml==4.4.2
nltk==3.4.5
beautifulsoup4==4.12.2
